// src/domain/enums/maintenance/MaintenanceStatus.ts
export enum MaintenanceStatus {
  REQUESTED = "requested",
  UNDER_EVALUATION = "under evaluation",
  APPROVED = "approved",
  ASSIGNED = "assigned",
  IN_PROGRESS = "in progress",
  PAUSED = "paused",
  COMPLETED = "completed",
  REJECTED = "rejected",
  CLOSED = "closed",
  CANCELED = "canceled",
}